+++ paq8pxd.cpp
@@ -133,6 +133,8 @@ inline int max(int a, int b) {return a<b?b:a;}
 #include "models/exe.hpp"
 #include "models/dec.hpp"
 #include "models/indirect.hpp"
+#include "models/record.hpp"
+#include "models/dmc.hpp"
 
 /////////////////////// Global context /////////////////////////
 U8 level=DEFAULT_OPTION;  // Compression level 0 to 15
@@ -250,7 +252,7 @@ const U32 tri[4]={0,4,3,7}, trj[4]={0,6,6,12};
 
 int dt[1024];  // i -> 16K/(i+i+3)
 
-int n0n1[256];
+int n0n1[256]; // for contectmap
 #define PNGFlag (1<<31)
 #define GrayFlag (1<<30)
 
@@ -386,10 +388,7 @@ inline bool CharInArray(const char c, const char a[], const int len) {
 }
 
 #define MAX_WORD_SIZE 64
-#define TAB 0x09
-#define NEW_LINE 0x0A
-#define CARRIAGE_RETURN 0x0D
-#define SPACE 0x20
+
 
 class Word {
 public:
@@ -3823,264 +3822,6 @@ public:
 
 
 
-//////////////////////////// recordModel ///////////////////////
-
-// Model 2-D data with fixed record length.  Also order 1-2 models
-// that include the distance to the last match.
-
-
-inline U8 Clip(int const Px){
-  if(Px>255)return 255;
-  if(Px<0)return 0;
-  return Px;
-}
-class recordModel1: public Model {
-   BlockData& x;
-   Buf& buf;
-   Array<int> cpos1, cpos2, cpos3, cpos4;
-   Array<int> wpos1; // buf(1..2) -> last position
-   Array<int> rlen;//, rlen1, rlen2, rlen3,rlenl;  // run length and 2 candidates
-   Array<int>  rcount;//, rcount2,rcount3;  // candidate counts
-    U8 padding; // detected padding byte
-   int prevTransition,nTransition  ; // position of the last padding transition
-   int col;
-   int mxCtx,x1;
-   bool MayBeImg24b;
-   ContextMap cm, cn, cq, co, cp;
-    int nMaps ;
-   StationaryMap Maps[6] ={ 10,10,8,8,8,{11,1} };
-    SmallStationaryContextMap sMap[3]{ {11, 1}, {3, 1}, {16,1} };
-    IndirectMap iMap[3]{ 8,8,8 };
-    IndirectContext<U16> iCtx[5]{ {16,8}, {16,8}, {16,8}, {20,8}, {11,1} };
-   U8 N, NN, NNN, NNNN,WxNW;
-   const int nIndCtxs  ;
-public:
-  recordModel1( BlockData& bd,U64 msize=CMlimit(MEM()*2) ): x(bd),buf(bd.buf),  cpos1(256) , cpos2(256),
-    cpos3(256), cpos4(256),wpos1(0x10000), rlen(3), rcount(2),padding(0),prevTransition(0),nTransition(0), col(0),mxCtx(0),
-    x1(0),MayBeImg24b(false),cm(32768, 3,M_RECORD), cn(32768/2, 3,M_RECORD),cq(32768*4, 3,M_RECORD), co(32768*2, 3,M_RECORD), cp(level>9?0x10000000 :CMlimit(msize*2), 16,M_RECORD), nMaps ( 6),
-    N(0), NN(0), NNN(0), NNNN(0),WxNW(0), nIndCtxs(5){
-        // run length and 2 candidates
-        rlen[0] = 2; 
-        rlen[1] = 0; 
-        rlen[2] = 0; 
-        // candidate counts
-        rcount[0] = 0;
-        rcount[1] = 0; 
-   }
-  int inputs() {return (3+3+3+3+16)*cm.inputs()+nMaps*2+2+3*2+3*2;}
-  int nets() {return 512+11*32+1024;}
-  int netcount() {return 2+1;}
-  int p(Mixer& m,int rrlen=0,int val2=0) {
-   // Find record length
-  if (!x.bpos) {
-   int w=x.c4&0xffff, c=w&255, d=w>>8,c1=x.b2,w1=(x.b3<<8)+x.b2,  e=x.c4&0xffffff;
-   bool dict=x.dictonline;
-    if (rrlen==0){
-        int r=0;
-        r=buf.pos-cpos1[c];
-      if ( r>1) {
-        if ( r==cpos1[c]-cpos2[c]  && r==cpos2[c]-cpos3[c] && (r>32 || r==cpos3[c]-cpos4[c])
-          && (r>10 || ((c==buf(r*5+1)) && c==buf(r*6+1)))) {      
-          if (r==rlen[1]) ++rcount[0];
-          else if (r==rlen[2]) ++rcount[1];
-          else if (rcount[0]>rcount[1]) rlen[2]=r, rcount[1]=1;
-          else rlen[1]=r, rcount[0]=1;
-        }
-      }    
-
-    // check candidate lengths
-    for (int i=0; i < 2; i++) {
-      if (rcount[i] > max(0,12-(int)ilog2(rlen[i+1]))){
-        if (rlen[0] != rlen[i+1]){
-            if (MayBeImg24b && rlen[i+1]==3){
-              rcount[0]>>=1;
-              rcount[1]>>=1;
-              continue;
-            }
-            else if ( (rlen[i+1] > rlen[0]) && (rlen[i+1] % rlen[0] == 0) ){
-            // maybe we found a multiple of the real record size..?
-            // in that case, it is probably an immediate multiple (2x).
-            // that is probably more likely the bigger the length, so
-            // check for small lengths too
-            if ((rlen[0] > 32) && (rlen[i+1] == rlen[0]*2)){
-              rcount[0]>>=1;
-              rcount[1]>>=1;
-              continue;
-            }
-          }
-          rlen[0] = rlen[i+1];
-          rcount[i] = 0;
-          MayBeImg24b = (rlen[0]>30 && (rlen[0]%3)==0);
-          nTransition = 0;
-        }
-        else
-          // we found the same length again, that's positive reinforcement that
-          // this really is the correct record size, so give it a little boost
-          rcount[i]>>=2;
-
-        // if the other candidate record length is orders of
-        // magnitude larger, it will probably never have enough time
-        // to increase its counter before it's reset again. and if
-        // this length is not a multiple of the other, than it might
-        // really be worthwhile to investigate it, so we won't set its
-        // counter to 0
-        if (rlen[i+1]<<4 > rlen[1+(i^1)])
-          rcount[i^1] = 0;
-      }
-    }
-    }
-    else rlen[0]=rrlen,rcount[0]=rcount[1]=0;
-    
-    assert(rlen[0]>0);
-    //if (dict==true)
-    //col=x.bufn.pos%x.wcol;
-    //else 
-    col=buf.pos%rlen[0];
-    x1 = min(0x1F,col/max(1,rlen[0]/32));
-    if (dict==true){
-        N = x.bufn(x.wcol&0xffff), NN = x.bufn((x.wcol*2)&0xffff), NNN = x.bufn((x.wcol*3)&0xffff), NNNN = x.bufn((x.wcol*4)&0xffff);
-    for (int i=0; i<nIndCtxs-1; iCtx[i]+=c, i++);
-    iCtx[0]=(c<<8)|N;
-    iCtx[1]=((x.wcol>1?x.bufn((x.wcol-1)&0xffff):0)<<8)|N;
-    iCtx[2]=(c<<8)|(x.wcol>1?x.bufn((x.wcol-1)&0xffff):0);
-    iCtx[3]=(hash(c, N, x.bufn((x.wcol+1)&0xffff)));
-    }else{
-    
-    N = buf(rlen[0]), NN = buf(rlen[0]*2), NNN = buf(rlen[0]*3), NNNN = buf(rlen[0]*4);
-    for (int i=0; i<nIndCtxs-1; iCtx[i]+=c, i++);
-    iCtx[0]=(c<<8)|N;
-    iCtx[1]=(buf(rlen[0]-1)<<8)|N;
-    iCtx[2]=(c<<8)|buf(rlen[0]-1);
-    iCtx[3]=(hash(c, N, buf(rlen[0]+1)));
-    }
-    /*
-    Consider record structures that include fixed-length strings.
-    These usually contain the text followed by either spaces or 0's,
-    depending on whether they're to be trimmed or they're null-terminated.
-    That means we can guess the length of the string field by looking
-    for small repetitions of one of these padding bytes followed by a
-    different byte. By storing the last position where this transition
-    ocurred, and what was the padding byte, we are able to model these
-    runs of padding bytes.
-    Special care is taken to skip record structures of less than 9 bytes,
-    since those may be little-endian 64 bit integers. If they contain
-    relatively low values (<2^40), we may consistently get runs of 3 or
-    even more 0's at the end of each record, and so we could assume that
-    to be the general case. But with integers, we can't be reasonably sure
-    that a number won't have 3 or more 0's just before a final non-zero MSB.
-    And with such simple structures, there's probably no need to be fancy
-    anyway
-    */
-    if (!col)
-      nTransition = 0;
-    if ((((U16)(x.c4>>8) == (U16)(SPACE*0x010101)) && (c != SPACE)) || (!(x.c4>>8) && c && ((padding != SPACE) || ((dict==true)?x.bufn.pos:buf.pos-prevTransition > rlen[0])))){
-      prevTransition = (dict==true)?x.bufn.pos:buf.pos;
-      nTransition+=(nTransition<31);
-      padding = (U8)d;
-    }
-
-    U64 i=0;
-    if (val2!=-1) {
-   
-    // Set 2 dimensional contexts
-    cm.set(hash(++i,c<<8| (min(255, (dict==true)?x.bufn.pos:buf.pos-cpos1[c])/4)));
-    if (!( x.filetype==DECA))
-    cm.set(hash(++i,w<<9| llog((dict==true)?x.bufn.pos:buf.pos-wpos1[w])>>2));
-    cm.set(hash(++i,rlen[0]|N<<10|NN<<18));
-
-    cn.set(hash(++i,w|rlen[0]<<8));
-    cn.set(hash(++i,d|rlen[0]<<16));
-    cn.set(hash(++i,c|rlen[0]<<8));
-
-    cq.set(hash(++i,w1<<3));
-    cq.set(hash(++i,c1<<19));
-    cq.set(hash(++i,e ));
-
-    co.set(hash(++i,buf(1)<<8|min(255, (dict==true)?x.bufn.pos:buf.pos-cpos1[buf(1)])));
-    co.set(hash(++i,buf(1)<<17|buf(2)<<9));
-    co.set(hash(++i,buf(1)<<8|N));
-    
-    cp.set(hash(++i,rlen[0]|N<<10|col<<18,x.wcol));
-    cp.set(hash(++i,rlen[0]|buf(1)<<10|col<<18,x.wcol));
-    cp.set(hash(++i,col|rlen[0]<<12,x.wcol));
-    if (rlen[0]>8){
-      cp.set( hash(++i, min(min(0xFF,rlen[0]),(dict==true)?x.bufn.pos:buf.pos-prevTransition), min(0x3FF,col), (w&0xF0F0)|(w==((padding<<8)|padding)), nTransition ) );
-      cp.set( hash(++i, w, ((dict==true)?x.bufn((x.wcol+1)&0xffff):buf(rlen[0]+1)==padding && N==padding), col/max(1,rlen[0]/32) ) );
-    }
-    else
-      cp.set(), cp.set();
-
-    cp.set(hash(++i, N|((NN&0xF0)<<4)|((NNN&0xE0)<<7)|((NNNN&0xE0)<<10)|((col/max(1,rlen[0]/16))<<18) ));
-    cp.set(hash(++i, (N&0xF8)|((NN&0xF8)<<8)|(col<<16),x.wcol ));
-    cp.set(hash(++i, N, NN));
-
-    cp.set(hash(++i, col, x.wcol,iCtx[0]()));
-    cp.set(hash(++i, col, x.wcol,iCtx[1]()));
-    cp.set(hash(++i, col, x.wcol,iCtx[0]()&0xFF, iCtx[1]()&0xFF));
-
-    cp.set(hash(++i, iCtx[2]()));  
-    cp.set(hash(++i, iCtx[3]()));
-    cp.set(hash(++i, iCtx[1]()&0xFF, iCtx[3]()&0xFF));
-
-    cp.set(hash(++i, N, (WxNW=c^(dict==true)?x.bufn((x.wcol+1)&0xffff):buf(rlen[0]+1))));
-    cp.set(hash(++i,(x.Match.length3!= 0) << 8U | x.Match.byte,U8(iCtx[1]()), N, WxNW));//cp.set(hash(++i, (min(ilog2(x.Match.length),3)!=0)<<8 |x.Match.byte,U8(iCtx[1]()), N, WxNW));
-    int k=0x300;
-    if (MayBeImg24b)
-      i = (col%3)<<8, Maps[0].set(Clip(((U8)(x.c4>>16))+c-(x.c4>>24))|k);
-    else
-      Maps[0].set(Clip(c*2-d)|k);
-    Maps[1].set(Clip(c+N-(dict==true)?x.bufn((x.wcol+1)&0xffff):buf(rlen[0]+1))|k);
-    Maps[2].set(Clip(N+NN-NNN));
-    Maps[3].set(Clip(N*2-NN));
-    Maps[4].set(Clip(N*3-NN*3+NNN));
-    iMap[0].set(N+NN-NNN);
-    iMap[1].set(N*2-NN);
-    iMap[2].set(N*3-NN*3+NNN);
-     }
-    // update last context positions
-    cpos4[c]=cpos3[c];
-    cpos3[c]=cpos2[c];
-    cpos2[c]=cpos1[c];
-    /*if (dict==true){
-    cpos1[c]=x.bufn.pos;
-    wpos1[w]=x.bufn.pos;
-    }else*/
-    {cpos1[c]=buf.pos;
-    wpos1[w]=buf.pos;
-    }
-    mxCtx = (rlen[0]>128)?(min(0x7F,col/max(1,rlen[0]/128))):col;
-  }
-  U8 B = x.c0<<(8-x.bpos);
-  U32 ctx = (N^B)|(x.bpos<<8);
-  iCtx[nIndCtxs-1]+=x.y, iCtx[nIndCtxs-1]=ctx;
-  Maps[nMaps-1].set(ctx);
-  sMap[0].set(ctx);
-  sMap[1].set(iCtx[nIndCtxs-1]());
-  sMap[2].set((ctx<<8)|WxNW);
-if (val2==-1) return rlen[0];
-  cm.mix(m);
-  cn.mix(m);
-  cq.mix(m);
-  co.mix(m);
-  cp.mix(m);
-  if (x.filetype==DICTTXT || x.filetype==BIGTEXT  || x.filetype==TEXT || x.filetype==TEXT0  ||x.filetype==EXE|| x.filetype==DECA || x.filetype==DBASE) {
-  }else{
-  
-  for (int i=0;i<nMaps;i++)
-    Maps[i].mix1(m, 1, 3);
-   for (int i=0;i<3;i++)
-    iMap[i].mix(m, 1, 3, 255); 
-    }
-  sMap[0].mix(m, 6, 1, 3);
-  sMap[1].mix(m, 6, 1, 3);
-  sMap[2].mix(m, 5, 1, 2);
-  m.set( (rlen[0] > 2) * ((x.bpos << 7U) | mxCtx), 1024);
-  m.set( ((N^B)>>4)|(x1<<4), 512 );
-  m.set( (x.grp<<5)|x1, 11*32);
-  return (rlen[0]>2)*( (x.bpos<<7)|mxCtx );//1024 
-  }
- virtual ~recordModel1(){ }
-};
 
 class linearPredictionModel: public Model {
   BlockData& x;
@@ -6963,256 +6704,6 @@ int p(Mixer& m,int info,int val2=0){
  
 };
 
-
-
-
-
-
-//////////////////////////// dmcModel //////////////////////////
-
-// Model using DMC (Dynamic Markov Compression).
-//
-// The bitwise context is represented by a state graph.
-//
-// See the original paper: http://webhome.cs.uvic.ca/~nigelh/Publications/DMC.pdf
-// See the original DMC implementation: http://maveric0.uwaterloo.ca/ftp/dmc/
-//
-// Main differences:
-// - Instead of floats we use fixed point arithmetic.
-// - For probability estimation each state maintains both a 0,1 count ("c0" and "c1") 
-//   and a bit history ("state"). The bit history is mapped to a probability adaptively using 
-//   a StateMap. The two computed probabilities are emitted to the Mixer to be combined.
-// - All counts are updated adaptively.
-// - The "dmcModel" is used in "dmcForest". See below.
-
-
-class dmcNode {
-private:
-struct DMCNode { // 12 bytes
-private:
-  // c0,c1: adaptive counts of zeroes and ones; 
-  //   fixed point numbers with 4 integer and 8 fractional bits, i.e. scaling factor=256;
-  //   thus the counts 0.0 .. 15.996 are represented by 0 .. 4095
-  // state: bit history state - as in a contextmodel
-  U32 state_c0_c1;  // 8 + 12 + 12 = 32 bits
-public:
-  U32 nx0,nx1;     //indexes of next DMC nodes in the state graph
-  U8   get_state() const {return state_c0_c1>>24;}
-  void set_state(U8 state) {state_c0_c1=(state_c0_c1 & 0x00FFFFFF)|(state<<24);}
-  U32 get_c0() const {return (state_c0_c1>>12) & 0xFFF;}
-  void set_c0(U32 c0) {assert(c0>=0 && c0<4096);state_c0_c1=(state_c0_c1 &0xFF000FFF) | (c0<<12);}
-  U32 get_c1() const {return state_c0_c1 & 0xFFF;}
-  void set_c1(U32 c1) {assert(c1>=0 && c1<4096);state_c0_c1=(state_c0_c1 &0xFFFFF000) | c1;}
-};
-  U32 top, curr;     // index of first unallocated node (i.e. number of allocated nodes); index of current node
-  U32 threshold;     // cloning threshold parameter: fixed point number as c0,c1
-  Array<DMCNode> t;  // state graph
-  StateMap sm;
-  BlockData& x;
-  // Initialize the state graph to a bytewise order 1 model
-  // See an explanation of the initial structure in:
-  // http://wing.comp.nus.edu.sg/~junping/docs/njp-icita2005.pdf
-  
-  void resetstategraph() {
-    assert(top==0 || top>65280);
-    for (int i=0; i<255; ++i) { //255 nodes in each tree
-      for (int j=0; j<256; ++j) { //256 trees
-        int node_idx=j*255+i;
-        if (i<127) { //internal tree nodes
-          t[node_idx].nx0=node_idx+i+1; // left node 
-          t[node_idx].nx1=node_idx+i+2; // right node
-        }
-        else { // 128 leaf nodes - they each references a root node of tree(i)
-          t[node_idx].nx0=(i-127)*255; // left node -> root of tree 0,1,2,3,... 
-          t[node_idx].nx1=(i+1)*255;   // right node -> root of tree 128,129,...
-        }
-        t[node_idx].set_c0(128); //0.5
-        t[node_idx].set_c1(128); //0.5
-        t[node_idx].set_state(0);
-      }
-    }
-    top=65280;
-    curr=0;
-  }
-
-  // helper function: adaptively increment a counter
-  U32 increment_counter (const U32 x, const U32 increment) const { //"*x" is a fixed point number as c0,c1 ; "increment"  is 0 or 1
-    return (((x<<4)-x)>>4)+(increment<<8); // x * (1-1/16) + increment*256
-  }
-
-  //update stategraph
-  void processbit(int y) {
-
-    U32 c0=t[curr].get_c0();
-    U32 c1=t[curr].get_c1();
-    const U32 n = y ==0 ? c0 : c1;
-
-    // update counts, state
-    t[curr].set_c0(increment_counter(c0,1-y));
-    t[curr].set_c1(increment_counter(c1,y));
-
-    t[curr].set_state(nex(t[curr].get_state(), y));
-
-    // clone next state when threshold is reached
-    const U32 next = y==0 ? t[curr].nx0 : t[curr].nx1;
-    c0=t[next].get_c0();
-    c1=t[next].get_c1();
-    const U32 nn=c0+c1;
-    if(n>=threshold && nn>=n+threshold && top<t.size()) {
-      U32 c0_top=U64(c0)*n/nn;
-      U32 c1_top=U64(c1)*n/nn;
-      assert(c0>=c0_top);
-      assert(c1>=c1_top);
-      c0-=c0_top;
-      c1-=c1_top;
-
-      t[top].set_c0(c0_top);
-      t[top].set_c1(c1_top);
-      t[next].set_c0(c0);
-      t[next].set_c1(c1);
-      
-      t[top].nx0=t[next].nx0;
-      t[top].nx1=t[next].nx1;
-      t[top].set_state(t[next].get_state());
-      if(y==0) t[curr].nx0=top;
-      else t[curr].nx1=top;
-      ++top;
-    }
-
-    if(y==0) curr=t[curr].nx0;
-    else     curr=t[curr].nx1;
-  }
-
-public: 
-  dmcNode(U32 mem, U32 th,BlockData& bd) : top(0),threshold(th),t(mem+(255*256)),  sm(),x(bd) {resetstategraph();  }//min(mem+(255*256),((U64(1)<<31)/sizeof(DMCNode)))
-
-  bool isfull() {return x.bpos==1 && top==t.size();}
-  bool isalmostfull() {return x.bpos==1 && top>=t.size()*15 >>4;} // *15/16
-  void reset() {resetstategraph();sm.Reset();}
-  void mix(Mixer& m, bool activate) {
-    processbit(m.x.y);
-    if(activate) {
-      const U32 n0=t[curr].get_c0()+1;
-      const U32 n1=t[curr].get_c1()+1;
-      const int pr1=(n1<<12)/(n0+n1);
-      const int pr2=sm.p(t[curr].get_state(),m.x.y);
-      m.add(stretch(pr1)>>2);
-      m.add(stretch(pr2)>>2);
-    }
-  }
-};
-
-// This class solves two problems of the DMC model
-// 1) The DMC model is a memory hungry algorighm. In theory it works best when it can clone
-//    nodes forever. But memory is a limited resource. When the state graph is full you can't
-//    clone nodes anymore. You can either i) reset the model (the state graph) and start over
-//    or ii) you can keep updating the counts forever in the already fixed state graph. Both
-//    choices are troublesome: i) resetting the model degrades the predictive power significantly
-//    until the graph becomes large enough again and ii) a fixed structure can't adapt anymore.
-//    To solve this issue:
-//    Two models with the same parameters work in tandem. Always both models are updated but
-//    only one model (the larger, mature one) is active (predicts) at any time. When one model
-//    needs resetting the other one feeds the mixer with predictions until the first one
-//    becomes mature (nearly full) again.
-//    Disadvantages: with the same memory reuirements we have just half of the number of nodes
-//    in each model. Also keeping two models updated at all times requires 2x as much
-//    calculations as updating one model only.
-//    Advantage: stable and better compression - even with reduced number of nodes.
-// 2) The DMC model is sensitive to the cloning threshold parameter. Some files prefer
-//    a smaller threshold other files prefer a larger threshold.
-//    The difference in terms of compression is significant.
-//    To solve this issue:
-//    Three models with different thresholds are used and their predictions are emitted to 
-//    the mixer. This way the model with the better threshold will be favored naturally.
-//    Disadvantage: same as in 1) just the available number of nodes is 1/3 of the 
-//    one-model case.
-//    Advantage: same as in 1).
-
-class dmcModel1: public Model {
-private:
-  U32 mem;
-  dmcNode dmcmodel1a; // models a and b have the same parameters and work in tandem
-  dmcNode dmcmodel1b;
-  dmcNode dmcmodel2a; // models 1,2,3 have different threshold parameters
-  dmcNode dmcmodel2b;
-  dmcNode dmcmodel3a;
-  dmcNode dmcmodel3b;
-  int model1_state=0; // initial state, model (a) is active, both models are growing
-  int model2_state=0; // model (a) is full and active, model (b) is reset and growing
-  int model3_state=0; // model (b) is full and active, model (a) is reset and growing
-  BlockData& x;
-public:
-  dmcModel1(BlockData& bd, U32 val=0):
-  mem(CMlimit(( (level<11?(0x10000UL<<level):((0x10000UL<<11))))/9)),
-  dmcmodel1a(mem,240,bd),
-  dmcmodel1b(mem,240,bd),
-  dmcmodel2a(mem,480,bd),
-  dmcmodel2b(mem,480,bd),
-  dmcmodel3a(mem,720,bd),
-  dmcmodel3b(mem,720,bd),
-  x(bd){}
-  int inputs() {return 2;}
-  int nets() {return 0;}
-  int netcount() {return 0;}
-  int p(Mixer& m,int val1=0,int val2=0){
-
-    switch(model1_state) {
-      case 0:
-        dmcmodel1a.mix(m,true);
-        dmcmodel1b.mix(m,false);
-        if(dmcmodel1a.isalmostfull()){dmcmodel1b.reset();model1_state++;}
-        break;
-      case 1:
-        dmcmodel1a.mix(m, true);
-        dmcmodel1b.mix(m, false);
-        if(dmcmodel1a.isfull() && dmcmodel1b.isalmostfull()){dmcmodel1a.reset();model1_state++;}
-        break;
-      case 2:
-        dmcmodel1b.mix(m,true);
-        dmcmodel1a.mix(m,false);
-        if(dmcmodel1b.isfull() && dmcmodel1a.isalmostfull()){dmcmodel1b.reset();model1_state--;}
-        break;
-    }
-    
-    switch(model2_state) {
-    case 0:
-      dmcmodel2a.mix(m,true);
-      dmcmodel2b.mix(m,false);
-      if(dmcmodel2a.isalmostfull()){dmcmodel2b.reset();model2_state++;}
-      break;
-    case 1:
-      dmcmodel2a.mix(m,true);
-      dmcmodel2b.mix(m,false);
-      if(dmcmodel2a.isfull() && dmcmodel2b.isalmostfull()){dmcmodel2a.reset();model2_state++;}
-      break;
-    case 2:
-      dmcmodel2b.mix(m,true);
-      dmcmodel2a.mix(m,false);
-      if(dmcmodel2b.isfull() && dmcmodel2a.isalmostfull()){dmcmodel2b.reset();model2_state--;}
-      break;
-    }
-
-    switch(model3_state) {
-    case 0:
-      dmcmodel3a.mix(m,true);
-      dmcmodel3b.mix(m,false);
-      if(dmcmodel3a.isalmostfull()){dmcmodel3b.reset();model3_state++;}
-      break;
-    case 1:
-      dmcmodel3a.mix(m,true);
-      dmcmodel3b.mix(m,false);
-      if(dmcmodel3a.isfull() && dmcmodel3b.isalmostfull()){dmcmodel3a.reset();model3_state++;}
-      break;
-    case 2:
-      dmcmodel3b.mix(m,true);
-      dmcmodel3a.mix(m,false);
-      if(dmcmodel3b.isfull() && dmcmodel3a.isalmostfull()){dmcmodel3b.reset();model3_state--;}
-      break;
-    }
-    return 0;
-  }
-  virtual ~dmcModel1(){ }
-};
 
 class nestModel1: public Model {
   BlockData& x;
